feat(chunkserver): Relocate hddDisksReinit
This change moves the hddDisksReinit function to the chunkserver-common
library, to make it available for the plugins. DiskManagers should
reimplement the virtual functions to provide custom behavior.

Driven by the movement, other related changes were applied:
- Rename the function to reloadDisksFromCfg which is more consistent.
- Move the gPluginManager to be visible from plugins, because it knows
  how to create the concrete disks. In this case, CmrDisk was needed
  as it is the default.
- The includes were also updated due to the movements.
- The DefaultDiskManager was also moved, to provide a base behavior.

The logic of the functions related to the disks loading was not
modified. They belong now to the current disk manager.
